+++ packages/tools/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./catalog.js";
 export * from "./filesystem.js";
 export * from "./policy.js";
+export * from "./search.js";
 export * from "./terminal.js";
 export * from "./web-fetch.js";